requirements.txt
django
psycopg
psycopg2-binary
djangorestframework
drf-yasg